requirements.txt
python-dotenv
pytenable
pandas
hvplot
jupyterlab
